Allow reading pass-through property of layers
PSD stores the pass-through property in the lsct section. After the 8
bits for the layer type, there may be another 8BIM magic number,
followed by either a repeated blend mode or the special "pass" mode,
which marks the group as being pass-through as opposed to isolated. Most
drawing programs let you manipulate this property directly, since it's
useful to have groups for organizational purposes, but not affecting the
way the layers inside get rendered.

This patch adds reading of this property. I don't think the exporter
supports layer groups, so there's no code for that added either.